Refactoring definition tree..
