+++ src/views/nodes/README.md
@@ -1,0 +1,6 @@
+# TODO
+
+1. Refactor node construction into providers
+1. Create providers through factory method
+1. Refactor node providers to get options from the caller instead of inheriting it from the nodes
+   - The options should come from the UI state only
